backend/requirements.txt
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0
python-dotenv==1.0.1
redis==5.1.0
httpx==0.27.0
google-generativeai>=0.8.0
pytesseract==0.3.13
Pillow==10.4.0
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
